requirements.txt
MouseInfo==0.1.3
Pillow==9.0.1
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyQt6==6.2.3
PyQt6-Qt6==6.2.3
PyQt6-sip==13.2.1
PyRect==0.1.4
PyScreeze==0.1.28
pytweening==1.0.4
